+++ src/char.cpp
@@ -504,7 +504,7 @@ void Character::clear_skills()
 
 int Character::get_skills_count() const
 {
-	return skills.size();
+	return static_cast<int>(skills.size());
 }
 
 int Character::get_obj_slot(int slot_num)
@@ -714,7 +714,7 @@ void change_fighting(CHAR_DATA * ch, int need_stop)
 	}
 }
 
-int fighting_list_size()
+size_t fighting_list_size()
 {
 	return fighting_list.size();
 }
@@ -1614,7 +1614,7 @@ std::string Character::race_or_title()
 	return result;
 }
 
-int Character::get_morphs_count() const
+size_t Character::get_morphs_count() const
 {
 	return morphs_.size();
 };
