+++ src/utils.cpp
@@ -272,7 +272,7 @@ int str_cmp(const std::string &arg1, const std::string &arg2)
  *
  * Scan until strings are found different, the end of both, or n is reached.
  */
-int strn_cmp(const char *arg1, const char *arg2, int n)
+int strn_cmp(const char *arg1, const char *arg2, size_t n)
 {
 	int chk, i;
 
@@ -288,7 +288,8 @@ int strn_cmp(const char *arg1, const char *arg2, int n)
 
 	return (0);
 }
-int strn_cmp(const std::string &arg1, const char *arg2, int n)
+
+int strn_cmp(const std::string &arg1, const char *arg2, size_t n)
 {
 	int chk;
 	std::string::size_type i;
@@ -311,7 +312,8 @@ int strn_cmp(const std::string &arg1, const char *arg2, int n)
 	else
 		return (LOWER(arg1[i]) - LOWER('\0'));
 }
-int strn_cmp(const char *arg1, const std::string &arg2, int n)
+
+int strn_cmp(const char *arg1, const std::string &arg2, size_t n)
 {
 	int chk;
 	std::string::size_type i;
@@ -334,7 +336,8 @@ int strn_cmp(const char *arg1, const std::string &arg2, int n)
 	else
 		return (LOWER('\0') - LOWER(arg2[i]));
 }
-int strn_cmp(const std::string &arg1, const std::string &arg2, int n)
+
+int strn_cmp(const std::string &arg1, const std::string &arg2, size_t n)
 {
 	int chk;
 	std::string::size_type i;
